Сделан урок 4, задание 1 (исправление)

+++ Lesson4/Task1/src/ru/x5/Shape.java
@@ -1,5 +1,5 @@
 package ru.x5;
 
 public interface Shape {
-    int square();
+    double square();
 }
